Update migration and self assessment scoring seeder

+++ database/migrations/2023_08_06_164652_create_self_assessment_score_components.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('self_assessment_score_components', function (Blueprint $table) {
             $table->id();
             $table->text('title')->nullable();
-            $table->integer('weight')->default(1);
+            $table->float('weight')->default(1);
         });
     }
